restructured the code: moved RevertionResult, CachedState and ItemState to a result package, all helper classes to the helper package, moved itemstack checking from AntiIllegals into RevertHelper, added more helper methods
