+++ src/main/java/org/zeroBzeroT/antiillegals/helpers/MaterialHelper.java
@@ -1,10 +1,16 @@
-package org.zeroBzeroT.antiillegals;
+package org.zeroBzeroT.antiillegals.helpers;
 
 import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
+import org.zeroBzeroT.antiillegals.MaterialSets;
 
-public class Checks {
+public class MaterialHelper {
+
+    private MaterialHelper() {
+
+    }
+
     public static boolean isIllegalBlock(@Nullable final Material material) {
         return material != null && MaterialSets.ILLEGAL_BLOCKS.contains(material);
     }
@@ -16,4 +22,5 @@ public class Checks {
     public static boolean isWeapon(@Nullable final ItemStack itemStack) {
         return itemStack != null && MaterialSets.WEAPON_MATERIALS.contains(itemStack.getType());
     }
+
 }
