+++ src/main/java/org/zeroBzeroT/antiillegals/result/RevertionResult.java
@@ -1,3 +1,3 @@
-package org.zeroBzeroT.antiillegals;
+package org.zeroBzeroT.antiillegals.result;
 
 public record RevertionResult(int books, boolean wasReverted) { }
